feat: authorize & onboard accepts `pkce` & `codeChallenge` options
`codeChallenge` - usually generated from a server from a code verifier,
pass this value to the authorize or onboard call and it redirects
back to the server where the token exchanging happens without
exposing the code verifier to the client side application.

`pkce` - this will tells the SDK to auto generate a code challenge,
this use case is for an application who wish to authorize or onboard
with PKCE flow and exchange the token on the client side.